requirements.txt
aniso8601==9.0.1
annotated-types==0.7.0
blinker==1.9.0
blis==1.0.1
catalogue==2.0.10
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
cloudpathlib==0.20.0
colorama==0.4.6
confection==0.1.5
cymem==2.0.8
elastic-transport==8.15.1
elasticsearch==8.16.0
Flask==3.1.0
Flask-Cors==5.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
langcodes==3.4.1
language_data==1.2.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
murmurhash==1.0.10
numpy==2.0.2
packaging==24.2
preshed==3.0.9
pydantic_core==2.23.4
Pygments==2.18.0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
rich==13.9.4
setuptools==75.3.0
shellingham==1.5.4
six==1.16.0
smart-open==7.0.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
thinc==8.3.2
typer==0.13.0
typing_extensions==4.12.2
urllib3==1.26.20
wasabi==1.1.3
weasel==0.4.1
Werkzeug==3.1.3
wrapt==1.16.0
